+++ src/shared/components/DataSetInput/DataSet3.tsx
@@ -1,5 +1,6 @@
-import { DataSet3 } from "../../../types/store";
-import Actions, { Action } from "../../store/actions";
+import { DataSet3 } from "../../../types/dataset";
+import { Action } from "../../../types/store/actions";
+import Actions from "../../store/actions";
 import Container from "./layouts/Container";
 import { DataSetInputLeftSide1 } from "./layouts/LeftSide1";
 import { DataSetInputRightSide } from "./layouts/RightSide1";
