+++ src/shared/store/dataSet1/reducer.ts
@@ -1,8 +1,8 @@
-import { State } from "../../../types/store";
+import { Action, EAction } from "../../../types/store/actions";
+import { State } from "../../../types/store/state";
 import { sanitizeText, textToArray } from "../../utils/convertions";
 import { newLabel } from "../../utils/newLabel";
 import { randomColor } from "../../utils/randomColor";
-import { Action, EAction } from "../actions";
 
 const Reducer = (state: State, action: Action): State => {
 	let clone = [...state.dataSet1];
