+++ src/shared/store/dataSet2/actions.ts
@@ -1,4 +1,4 @@
-import { Action, EAction } from "../actions";
+import { Action, EAction } from "../../../types/store/actions";
 
 export const dataSet2 = {
 	clear: (): Action => ({ type: EAction.DATASET2_CLEAR, payload: null }),
