+++ src/shared/store/dataSet3/reducer.ts
@@ -1,7 +1,7 @@
-import { State } from "../../../types/store";
+import { Action, EAction } from "../../../types/store/actions";
+import { State } from "../../../types/store/state";
 import { sanitizeText, textToArray } from "../../utils/convertions";
 import { randomColor } from "../../utils/randomColor";
-import { Action, EAction } from "../actions";
 
 const Reducer = (state: State, action: Action): State => {
 	switch (action.type) {
