+++ src/shared/utils/chartAdapters/adapter5.ts
@@ -1,4 +1,4 @@
-import { DataSet3 } from "../../../types/store";
+import { DataSet3 } from "../../../types/dataset";
 
 export const adapter5 = (input: DataSet3) => {
 	if (input.array.length === 0) {
